Close #30 - Extract the essential properties of `Refined` and create `RefinedBase`
